PRSD-547: Move classes to appropriate packages
